Added PCOThemeWindow
This window listens for theme changes.  When the theme changes it will take a snapshot of its self and after all the theme changes have been processed then it will fade out the snapshot view and transition to the new UI.

Setting the `shouldAnimateThemeChange` property to NO will skit the snapshot and run through the theme changes the same.

Unlike UIResponders that need to start and stop listening for theme changes, PCOThemeWindow will automatically start and stop listening for theme changes.  The reason for this behavior and not manual start/stop behavior is because it uses different notifications to process the cages.

PCOThemeView can still listen for the regular theme changes and there is no special requirements for starting/stoping outside of the normal behavior.